refactor(livestock): (AHWR-2044) rename files to match slugs #patch
