+++ app/constants/routes.js
@@ -42,10 +42,10 @@ export const poultryApplyRoutes = {
 };
 
 export const applyViews = {
-  youCanClaimMultiple: "livestock/apply/you-can-claim-multiple",
+  youCanClaimMultiple: "livestock/apply/what-you-can-claim",
   timings: "livestock/apply/timings",
-  numbers: "livestock/apply/numbers",
-  declaration: "livestock/apply/declaration",
+  numbers: "livestock/apply/minimum-number",
+  declaration: "livestock/apply/agreement-offer",
   offerRejected: "livestock/apply/offer-rejected",
   confirmation: "livestock/apply/confirmation",
 };
@@ -103,26 +103,26 @@ export const claimRoutes = {
 export const claimViews = {
   assuranceScheme: "livestock/claim/assurance-scheme",
   biosecurityException: "livestock/claim/biosecurity-exception",
-  biosecurity: "livestock/claim/biosecurity",
+  biosecurity: "livestock/claim/biosecurity-assessment",
   checkAnswers: "livestock/claim/check-answers",
   checkHerdDetails: "livestock/claim/check-herd-details",
-  confirmation: "livestock/claim/confirmation",
+  confirmation: "livestock/claim/claim-confirmation",
   dairyFollowUpDateException: "livestock/claim/dairy-follow-up-date-exception",
   dateOfTestingException: "livestock/claim/date-of-testing-exception",
-  dateOfTesting: "livestock/claim/date-of-testing",
+  dateOfTesting: "livestock/claim/test-date",
   dateOfVisitException: "livestock/claim/date-of-visit-exception",
   dateOfVisit: "livestock/claim/date-of-visit",
   diseaseStatus: "livestock/claim/disease-status",
-  enterCphNumber: "livestock/claim/enter-cph-number",
+  enterCphNumber: "livestock/claim/cph",
   enterHerdDetails: "livestock/claim/enter-herd-details",
-  enterHerdName: "livestock/claim/enter-herd-name",
-  herdOthersOnSbi: "livestock/claim/herd-others-on-sbi",
+  enterHerdName: "livestock/claim/herd-name",
+  herdOthersOnSbi: "livestock/claim/sbi-herds",
   multipleSpeciesDateException: "livestock/claim/multiple-species-date-exception",
-  typeOfSamplesTaken: "livestock/claim/type-of-samples-taken",
+  typeOfSamplesTaken: "livestock/claim/samples-types",
   numberOfFluidOralSamplesException: "livestock/claim/number-of-fluid-oral-samples-exception",
-  numberOfFluidOralSamples: "livestock/claim/number-of-fluid-oral-samples",
+  numberOfFluidOralSamples: "livestock/claim/oral-sample-number",
   numberOfBloodSamplesException: "livestock/claim/number-of-blood-samples-exception",
-  numberOfBloodSamples: "livestock/claim/number-of-blood-samples",
+  numberOfBloodSamples: "livestock/claim/blood-sample-number",
   numberOfSamplesTestedException: "livestock/claim/number-of-samples-tested-exception",
   numberOfSamplesTested: "livestock/claim/number-of-samples-tested",
   numberOfSpeciesException: "livestock/claim/number-of-species-exception",
@@ -142,7 +142,7 @@ export const claimViews = {
   sameHerd: "livestock/claim/same-herd",
   selectTheHerdDateException: "livestock/claim/select-the-herd-date-exception",
   selectTheHerdException: "livestock/claim/select-the-herd-exception",
-  selectTheHerd: "livestock/claim/select-the-herd",
+  selectTheHerd: "livestock/claim/select-herd",
   sheepEndemicsPackage: "livestock/claim/sheep-endemics-package",
   sheepTestResults: "livestock/claim/sheep-test-results",
   sheepTests: "livestock/claim/sheep-tests",
@@ -155,11 +155,11 @@ export const claimViews = {
   vetName: "livestock/claim/vet-name",
   vetRcvs: "livestock/claim/vet-rcvs",
   vetVisitsReviewTestResults: "livestock/claim/vet-visits-review-test-results",
-  whichSpecies: "livestock/claim/which-species",
+  whichSpecies: "livestock/claim/species",
   whichTypeOfReviewDairyFollowUpException:
     "livestock/claim/which-type-of-review-dairy-follow-up-exception",
   whichTypeOfReviewException: "livestock/claim/which-type-of-review-exception",
-  whichTypeOfReview: "livestock/claim/which-type-of-review",
+  whichTypeOfReview: "livestock/claim/review-type",
   youCannotClaim: "livestock/claim/you-cannot-claim",
 };
