+++ app/plugins/router.js
@@ -5,7 +5,7 @@ import { entryPointHandlers } from "../routes/index.js";
 import { updateDetailsHandlers } from "../routes/update-details.js";
 import { signinRouteHandlers } from "../routes/signin-oidc.js";
 import { downloadApplicationHandlers } from "../routes/download-application.js";
-import { vetVisitsHandlers } from "../routes/livestock/vet-visits.js";
+import { vetVisitsHandlers } from "../routes/livestock/manage-claims.js";
 import { devLoginHandlers } from "../routes/dev-landing-page.js";
 import { accessibilityRouteHandlers } from "../routes/accessibility.js";
 import { signOutHandlers } from "../routes/sign-out.js";
@@ -13,30 +13,30 @@ import { config } from "../config/index.js";
 import { cannotSignInExceptionHandlers } from "../routes/cannot-sign-in.js";
 import { defraIdSignInHandlers } from "../routes/sign-in.js";
 import { missingPagesRoutes } from "../routes/missing-routes.js";
-import { declarationRouteHandlers } from "../routes/livestock/apply/declaration.js";
-import { numbersRouteHandlers } from "../routes/livestock/apply/numbers.js";
+import { declarationRouteHandlers } from "../routes/livestock/apply/agreement-offer.js";
+import { numbersRouteHandlers } from "../routes/livestock/apply/minimum-number.js";
 import { timingsRouteHandlers } from "../routes/livestock/apply/timings.js";
-import { claimMultipleRouteHandlers } from "../routes/livestock/apply/you-can-claim-multiple.js";
+import { claimMultipleRouteHandlers } from "../routes/livestock/apply/what-you-can-claim.js";
 import { poultryDeclarationRouteHandlers } from "../routes/poultry/apply/agreement-offer.js";
 import { poultryNumbersRouteHandlers } from "../routes/poultry/apply/minimum-number.js";
 import { poultryTimingsRouteHandlers } from "../routes/poultry/apply/timings.js";
 import { poultryClaimMultipleRouteHandlers } from "../routes/poultry/apply/what-you-can-claim.js";
 import { poultryDateOfVisitHandlers } from "../routes/poultry/claim/date-of-visit.js";
 import { selectFundingRouteHandlers } from "../routes/select-funding.js";
-import { biosecurityHandlers } from "../routes/livestock/claim/biosecurity.js";
+import { biosecurityHandlers } from "../routes/livestock/claim/biosecurity-assessment.js";
 import { checkAnswersHandlers } from "../routes/livestock/claim/check-answers.js";
 import { checkHerdDetailsHandlers } from "../routes/livestock/claim/check-herd-details.js";
-import { confirmationHandlers } from "../routes/livestock/claim/confirmation.js";
-import { dateOfTestingHandlers } from "../routes/livestock/claim/date-of-testing.js";
+import { confirmationHandlers } from "../routes/livestock/claim/claim-confirmation.js";
+import { dateOfTestingHandlers } from "../routes/livestock/claim/test-date.js";
 import { dateOfVisitHandlers } from "../routes/livestock/claim/date-of-visit.js";
 import { diseaseStatusHandlers } from "../routes/livestock/claim/disease-status.js";
-import { enterCphNumberHandlers } from "../routes/livestock/claim/enter-cph-number.js";
+import { enterCphNumberHandlers } from "../routes/livestock/claim/cph.js";
 import { enterHerdDetailsHandlers } from "../routes/livestock/claim/enter-herd-details.js";
-import { enterHerdNameHandlers } from "../routes/livestock/claim/enter-herd-name.js";
-import { herdOthersOnSbiHandlers } from "../routes/livestock/claim/herd-others-on-sbi.js";
-import { typeOfSamplesTakenHandlers } from "../routes/livestock/claim/type-of-samples-taken.js";
-import { numberOfOralFluidSamplesHandlers } from "../routes/livestock/claim/number-of-fluid-oral-samples.js";
-import { numberOfBloodSamplesHandlers } from "../routes/livestock/claim/number-of-blood-samples.js";
+import { enterHerdNameHandlers } from "../routes/livestock/claim/herd-name.js";
+import { herdOthersOnSbiHandlers } from "../routes/livestock/claim/sbi-herds.js";
+import { typeOfSamplesTakenHandlers } from "../routes/livestock/claim/samples-types.js";
+import { numberOfOralFluidSamplesHandlers } from "../routes/livestock/claim/oral-sample-number.js";
+import { numberOfBloodSamplesHandlers } from "../routes/livestock/claim/blood-sample-number.js";
 import { numberOfSamplesTestedHandlers } from "../routes/livestock/claim/number-of-samples-tested.js";
 import { numberOfSpeciesTestedHandlers } from "../routes/livestock/claim/number-of-species-tested.js";
 import { piHuntHandlers } from "../routes/livestock/claim/pi-hunt.js";
@@ -46,7 +46,7 @@ import { pigsElisaResultHandlers } from "../routes/livestock/claim/pigs-elisa-re
 import { pigsGeneticSequencingHandlers } from "../routes/livestock/claim/pigs-genetic-sequencing.js";
 import { pigsPcrResultHandlers } from "../routes/livestock/claim/pigs-pcr-result.js";
 import { sameHerdHandlers } from "../routes/livestock/claim/same-herd.js";
-import { selectTheHerdHandlers } from "../routes/livestock/claim/select-the-herd.js";
+import { selectTheHerdHandlers } from "../routes/livestock/claim/select-herd.js";
 import { sheepEndemicsPackageHandlers } from "../routes/livestock/claim/sheep-endemics-package.js";
 import { sheepTestResultsHandlers } from "../routes/livestock/claim/sheep-test-results.js";
 import { sheepTestsHandlers } from "../routes/livestock/claim/sheep-tests.js";
@@ -57,8 +57,8 @@ import { vaccinationHandlers } from "../routes/livestock/claim/vaccination.js";
 import { vetsNameHandlers } from "../routes/livestock/claim/vet-name.js";
 import { vetRCVSHandlers } from "../routes/livestock/claim/vet-rcvs.js";
 import { vetVisitsReviewTestResultsHandlers } from "../routes/livestock/claim/vet-visits-review-test-results.js";
-import { whichSpeciesHandlers } from "../routes/livestock/claim/which-species.js";
-import { whichReviewHandlers } from "../routes/livestock/claim/which-type-of-review.js";
+import { whichSpeciesHandlers } from "../routes/livestock/claim/species.js";
+import { whichReviewHandlers } from "../routes/livestock/claim/review-type.js";
 import { assuranceSchemeHandlers } from "../routes/livestock/claim/assurance-scheme.js";
 import { poultryVetVisitsHandlers } from "../routes/poultry/manage-claims.js";
 import { poultrySelectTheSiteHandlers } from "../routes/poultry/claim/select-the-site.js";
