+++ app/views/livestock/claim/samples-types.njk
